+++ src/lib/components/common/jschema/EnumProperty.svelte
@@ -7,14 +7,17 @@
 	/** @type {import('$lib/components/common/jschema/schema_management').SchemaProperty} */
 	export let schemaProperty;
 
-	/** @type {HTMLInputElement} */
+	function handleValueChange() {
+		schemaManager.updateValue(schemaProperty.key, schemaProperty.value);
+	}
+
+	/** @type {HTMLSelectElement} */
 	let field;
 	let validationError = '';
 
-	function handleValueChange() {
-		schemaManager.updateValue(schemaProperty.key, schemaProperty.value);
+	function validate() {
 		validationError = '';
-		if (schemaProperty.isRequired() && field.value === '') {
+		if (schemaProperty.isRequired() && (!field.value || field.value === 'null')) {
 			validationError = 'Field is required';
 		}
 	}
@@ -23,20 +26,27 @@
 <div class="d-flex align-items-center p-2">
 	<div class="property-metadata d-flex flex-row align-self-center w-50">
 		<label class={schemaProperty.isRequired() ? 'fw-bold' : ''} for="property-{schemaProperty.key}">
-			{schemaProperty.title || 'String argument'}
+			{schemaProperty.title || 'Enum argument'}
 		</label>
 		<PropertyDescription description={schemaProperty.description} />
 	</div>
 	<div class="property-input ms-auto w-50 has-validation">
-		<input
-			type="text"
+		<select
 			bind:this={field}
 			bind:value={schemaProperty.value}
-			on:input={handleValueChange}
+			on:change={handleValueChange}
+			on:input={validate}
 			class="form-control"
 			id="property-{schemaProperty.key}"
 			class:is-invalid={validationError}
-		/>
+		>
+			<option value={null}>Select...</option>
+			{#if schemaProperty.enum}
+				{#each schemaProperty.enum as optionValue}
+					<option>{optionValue}</option>
+				{/each}
+			{/if}
+		</select>
 		<span class="invalid-feedback">{validationError}</span>
 	</div>
 </div>